Move 'ExternalFile' to asdf and update changelog
